+++ ViewModel/CustomControls/NotificationBar/NotificationBarNotification.cs
@@ -117,7 +117,7 @@ namespace GolemUI.ViewModel.CustomControls
             }
         }
 
-        public event PropertyChangedEventHandler PropertyChanged;
+        public event PropertyChangedEventHandler? PropertyChanged;
 
         private void OnPropertyChanged(string? propertyName)
         {
