+++ ViewModel/DashboardViewModel.cs
@@ -8,16 +8,12 @@ using System.Diagnostics;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using System.Windows;
-using System.Windows.Controls;
-using System.Windows.Media;
-using System.Windows.Media.Animation;
 
 namespace GolemUI.ViewModel
 {
-    public class DashboardViewModel : INotifyPropertyChanged
+    public partial class DashboardViewModel : INotifyPropertyChanged
     {
-        public DashboardViewModel(DashboardMain dashboardMain, DashboardSettings dashboardSettings, DashboardSettingsAdv dashboardSettingsAdv, DashboardWallet dashboardWallet, NotificationBar notificationBar)
+        public DashboardViewModel(DashboardMain dashboardMain, DashboardSettings dashboardSettings, DashboardSettingsAdv dashboardSettingsAdv, DashboardWallet dashboardWallet)
         {
 
             PropertyChanged += OnPropertyChanged;
@@ -48,129 +44,10 @@ namespace GolemUI.ViewModel
             PageDashboardStatistics
         }
 
-        public class DashboardPage
-        {
-            public UserControl View;
-            ISavableLoadableDashboardPage? ViewModel = null;
-            public bool ShouldAutoLoad = false;
-            public bool ShouldAutoSave = false;
-            bool ShouldAnimate = true;
-            public DashboardPage(UserControl view)
-            {
-                View = view;
-                view.Visibility = Visibility.Hidden;
-                view.Opacity = 0;
-            }
 
-            public event PageChangeRequestedEvent? PageChangeRequested;
-            public event RequestDarkBackgroundEventHandler? DarkBackgroundRequested;
-            public DashboardPage(UserControl view, object viewModel)
-            {
-                View = view;
-                if (viewModel is ISavableLoadableDashboardPage model)
-                {
-                    ViewModel = model;
-                    ShouldAutoLoad = true;
-                    ShouldAutoSave = true;
-                    view.Visibility = Visibility.Hidden;
-                    view.Opacity = 0;
-                    model.PageChangeRequested += ViewModel_PageChangeRequested;
-                }
-
-                if (viewModel is IDialogInvoker dialogInvoker)
-                {
-                    dialogInvoker.DarkBackgroundRequested += DialogInvoker_DarkBackgroundRequested;
-                }
-            }
-
-            private void DialogInvoker_DarkBackgroundRequested(bool shouldBackgroundBeVisible)
-            {
-                DarkBackgroundRequested?.Invoke(shouldBackgroundBeVisible);
-            }
-
-
-            private void ViewModel_PageChangeRequested(DashboardViewModel.DashboardPages page)
-            {
-                PageChangeRequested?.Invoke(page);
-            }
-
-            public void Unmount()
-            {
-                if (ShouldAutoSave && ViewModel != null)
-                {
-                    ViewModel.SaveData();
-                }
-            }
-
-            public void Mount()
-            {
-                if (ShouldAutoLoad && ViewModel != null)
-                {
-                    ViewModel.LoadData();
-                }
-            }
-            public void Show()
-            {
-                if (ShouldAnimate)
-                {
-                    View.Visibility = Visibility.Visible;
-                    ShowSlowly(TimeSpan.FromMilliseconds(250));
-                }
-                else
-                {
-                    View.Visibility = Visibility.Visible;
-                    View.Opacity = 1.0f;
-                }
-            }
-
-            public void Hide()
-            {
-                if (ShouldAnimate)
-                {
-                    HideSlowly(TimeSpan.FromMilliseconds(300));
-                }
-                else
-                {
-                    View.Opacity = 0.0f;
-                    View.Visibility = Visibility.Hidden;
-                }
-            }
-            public void Clear()
-            {
-                View.Visibility = Visibility.Hidden;
-                View.Opacity = 0.0f;
-            }
-            private void ShowSlowly(TimeSpan duration)
-            {
-
-                DoubleAnimation animation = new DoubleAnimation(0.0, 1.0, duration);
-                DoubleAnimation animation2 = new DoubleAnimation(0.92, 1.0, duration);
-
-                View.RenderTransform = new ScaleTransform(0.9, 0.9, 0.5, 0.5);
-                View.RenderTransform.BeginAnimation(ScaleTransform.ScaleYProperty, animation2);
-                View.RenderTransform.BeginAnimation(ScaleTransform.ScaleXProperty, animation2);
-                View.BeginAnimation(UserControl.OpacityProperty, animation);
-
-            }
-            private void HideSlowly(TimeSpan duration)
-            {
-                DoubleAnimation anim1 = new DoubleAnimation(1, 0, duration);
-                anim1.Completed += new EventHandler(delegate (object? s, EventArgs ev)
-                {
-                    View.Visibility = Visibility.Hidden;
-                });
-                View.BeginAnimation(UserControl.OpacityProperty, anim1);
-
-                /*DoubleAnimation animation = new DoubleAnimation(1.0, 0.0, duration);
-                uc.BeginAnimation(UserControl.OpacityProperty, animation);*/
-            }
-        }
-
-
-        public Dictionary<DashboardPages, DashboardPage> _pages = new Dictionary<DashboardPages, DashboardPage>();
+        public Dictionary<DashboardPages, DashboardPage> _pages = new();
 
         public DashboardMain DashboardMain { get; set; }
-        public NotificationBar NotificationBar { get; set; }
         public DashboardSettings DashboardSettings { get; set; }
         public DashboardSettingsAdv DashboardSettingsAdv { get; set; }
         public DashboardWallet DashboardWallet { get; set; }
@@ -185,21 +62,15 @@ namespace GolemUI.ViewModel
         {
             get
             {
-                switch (_selectedPage)
+                return _selectedPage switch
                 {
-                    case DashboardPages.PageDashboardMain:
-                        return 1;
-                    case DashboardPages.PageDashboardSettings:
-                        return 2;
-                    case DashboardPages.PageDashboardSettingsAdv:
-                        return 2;
-                    case DashboardPages.PageDashboardWallet:
-                        return 3;
-                    case DashboardPages.PageDashboardStatistics:
-                        return 4;
-                    default:
-                        return -1;
-                }
+                    DashboardPages.PageDashboardMain => 1,
+                    DashboardPages.PageDashboardSettings => 2,
+                    DashboardPages.PageDashboardSettingsAdv => 2,
+                    DashboardPages.PageDashboardWallet => 3,
+                    DashboardPages.PageDashboardStatistics => 4,
+                    _ => -1,
+                };
             }
             set
             {
@@ -223,7 +94,7 @@ namespace GolemUI.ViewModel
                 {
                     Debug.Fail("No such page");
                 }
-                OnPropertyChanged("SelectedPage");
+                OnPropertyChanged(nameof(SelectedPage));
             }
         }
 
@@ -233,7 +104,7 @@ namespace GolemUI.ViewModel
             get => _darkBackgroundVisible; set
             {
                 _darkBackgroundVisible = value;
-                OnPropertyChanged("DarkBackgroundVisible");
+                OnPropertyChanged(nameof(DarkBackgroundVisible));
             }
         }
 
@@ -275,10 +146,7 @@ namespace GolemUI.ViewModel
 
         private void OnPropertyChanged(string? propertyName)
         {
-            if (PropertyChanged != null)
-            {
-                PropertyChanged(this, new PropertyChangedEventArgs(propertyName));
-            }
+            PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(propertyName));
         }
 
         private void OnPropertyChanged(object sender, PropertyChangedEventArgs e)
